+++ src/main/java/hello/AuthControllerAdvice.java
@@ -8,7 +8,7 @@ import org.springframework.web.bind.annotation.ControllerAdvice;
 import hello.repositories.UserRepository;
 import hello.entities.AppUser;
 import java.util.List;
-
+import hello.services.MembershipService;
 
 @ControllerAdvice
 public class AuthControllerAdvice {
